nc/requirements.txt
asgiref==3.7.2
asttokens==2.4.1
black==23.12.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.1
decorator==5.1.1
defusedxml==0.8.0rc2
Django==4.2.7
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
executing==2.0.1
gunicorn==21.0.1
hiredis==2.3.2
idna==3.6
ipython==8.18.1
jedi==0.19.1
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
python3-openid==3.2.0
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.0
tzdata==2023.3
urllib3==2.1.0
uuid==1.30
wcwidth==0.2.12
